Light copyedit
I tweaked the boilerplate text at the beginning, which is something I've been meaning to suggest for all the "Try it at home!" recipes. Let me know what you think. I also assumed that these should not be in code format: "process_name" and "process_command_line," so I left them as is.